Remove unnecessary code, update launch settings, add logo
