+++ .gitignore
@@ -1,7 +1,6 @@
 out
 dist
 node_modules
-/.vscode
 .vscode-test/
 *.vsix
 .DS_Store
