--- src/definitions/QExtension.ts
@@ -1,7 +1,0 @@
-export interface QExtension {
-  name: string;
-  labels: string[];
-  groupId: string;
-  artifactId: string;
-  guide?: string;
-}
