+++ src/definitions/qExtension.ts
@@ -1,0 +1,7 @@
+export interface QExtension {
+  name: string;
+  labels: string[];
+  groupId: string;
+  artifactId: string;
+  guide?: string;
+}
