+++ src/generateProject/pickExtensions.ts
@@ -1,7 +1,7 @@
 import { MultiStepInput } from '../utils/multiStepUtils';
-import { QExtension } from '../definitions/QExtension';
-import { State } from '../definitions/State';
-import { SettingsJson } from '../definitions/ConfigManager';
+import { QExtension } from '../definitions/qExtension';
+import { State } from '../definitions/state';
+import { SettingsJson } from '../definitions/configManager';
 import { getQExtensions } from '../utils/requestUtils';
 import { DEFAULT_API_URL } from '../definitions/constants';
